corrected data model + added first data to load at startup of the app

+++ src/main/java/com/account/manager/am/controller/BackendController.java
@@ -20,7 +20,7 @@ public class BackendController {
     }
 
     @GetMapping("/customer/{customerId}")
-    public Customer getCustomer(@PathVariable Long customerId) {
+    public Customer getCustomer(@PathVariable int customerId) {
         return customerService.getCustomerById(customerId);
     }
 }
