+++ src/main/java/com/account/manager/am/service/CustomerService.java
@@ -4,5 +4,5 @@ import com.account.manager.am.model.Customer;
 
 public interface CustomerService {
 
-    Customer getCustomerById(Long customerId);
+    Customer getCustomerById(int customerId);
 }
